+++ week3-task/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
+    "axios": "^1.9.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
   },
